Split IndexRegionsTest and UpdateRegionMappingTest into ES5 and ES8 variants
ES5 tests moved to ElasticSearch5/ subfolders and implement the
ElasticSearch5Test marker interface. ES8 tests stay in place with
the _on_es8 suffix dropped from test method names. UpdateRegionMappingTest
now extends AbstractOperationTestCase directly instead of AbstractMappingTestCase
since the latter is ES5-only.
