+++ tests/ElasticSearch/Operations/ElasticSearch5/UpdateRegionMappingTest.php
@@ -2,19 +2,27 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Search\ElasticSearch\Operations;
+namespace CultuurNet\UDB3\Search\ElasticSearch\Operations\ElasticSearch5;
 
+use CultuurNet\UDB3\Search\ElasticSearch\Operations\AbstractMappingTestCase;
+use CultuurNet\UDB3\Search\ElasticSearch\Operations\UpdateRegionMapping;
+use CultuurNet\UDB3\Search\ElasticSearch5Test;
 use Elasticsearch\Client;
 use Psr\Log\LoggerInterface;
 
-final class UpdateRegionMappingTest extends AbstractOperationTestCase
+final class UpdateRegionMappingTest extends AbstractMappingTestCase implements ElasticSearch5Test
 {
     protected function createOperation(Client $client, LoggerInterface $logger): UpdateRegionMapping
     {
         return new UpdateRegionMapping($client, $logger);
     }
 
-    private function getExpectedMappingBody(): array
+    protected function getDocumentType(): string
+    {
+        return 'region';
+    }
+
+    protected function getExpectedMappingBody(): array
     {
         return [
             'properties' => [
@@ -25,18 +33,24 @@ final class UpdateRegionMappingTest extends AbstractOperationTestCase
         ];
     }
 
+    protected function runOperation(string $indexName): void
+    {
+        $this->operation->run($indexName, $this->getDocumentType());
+    }
+
     /**
      * @test
      */
-    public function it_updates_the_mapping_without_type(): void
+    public function it_updates_the_mapping_with_type_on_es5(): void
     {
         $indexName = 'mock';
-        $documentType = 'region';
+        $documentType = $this->getDocumentType();
 
         $this->indices->expects($this->once())
             ->method('putMapping')
             ->with([
                 'index' => $indexName,
+                'type' => $documentType,
                 'body' => $this->getExpectedMappingBody(),
             ]);
 
@@ -44,6 +58,7 @@ final class UpdateRegionMappingTest extends AbstractOperationTestCase
             ->method('info')
             ->with("Mapping for type {$documentType} updated.");
 
+        $this->operation->enableElasticSearch5CompatibilityMode();
         $this->operation->run($indexName, $documentType);
     }
 }
